pip-requirements.txt
## @file
# PIP requirements file for Github Action Check CodeOwners and Maintainers
#
# This file provides the list of python components to install using PIP.
#
# Copyright (c) 2022, Intel Corporation. All rights reserved.<BR>
# SPDX-License-Identifier: BSD-2-Clause-Patent
#
# https://pypi.org/project/pip/
# https://pip.pypa.io/en/stable/user_guide/#requirements-files
# https://pip.pypa.io/en/stable/reference/pip_install/#requirements-file-format
# https://www.python.org/dev/peps/pep-0440/#version-specifiers
##

codeowners==0.5.0
GitPython==3.1.14
cached-property==1.5.2
